# Make header parsing "safe"