… src/Servers/Kestrel/Core/src/Internal/Http/HttpProtocol.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpRequestHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Infrastructure/HttpUtilities.cs
… src/Servers/Kestrel/Core/test/HttpParserTests.cs
… src/Servers/Kestrel/shared/test/HttpParsingData.cs
… src/Shared/ServerInfrastructure/BufferExtensions.cs