@benaadams 2020-04-06
Meh skipping bounds check for one space doesn't make a great difference and doesn't look very pretty; so reverted