@benaadams 2020-04-08
Reverted last commit as performance wasn't good.

Should be ready to review