@davidfowl 2020-04-08
@benaadams microbenchmarks?