@adamsitnik 2020-04-08
I've fetched these changes and run them against the [microbenchmarks](https://github.com/adamsitnik/aspnetcore/blob/techEmpowerParsing/src/Servers/Kestrel/perf/Kestrel.Performance/TechEmpowerHttpParserBenchmark.cs) from https://github.com/dotnet/aspnetcore/issues/20518

Before:

|               Method |       Mean |     Error |    StdDev |        Op/s |
|--------------------- |-----------:|----------:|----------:|------------:|
| PlaintextTechEmpower | 4,214.6 ns | 81.241 ns | 72.018 ns |   237,268.2 |
|      JsonTechEmpower |   271.9 ns |  1.483 ns |  1.314 ns | 3,677,397.6 |

After:

|               Method |       Mean |    Error |   StdDev |        Op/s |
|--------------------- |-----------:|---------:|---------:|------------:|
| PlaintextTechEmpower | 3,766.7 ns |  4.94 ns |  3.86 ns |   265,486.6 |
|      JsonTechEmpower |   252.5 ns |  0.55 ns |  0.48 ns | 3,960,332.8 |

I've also run it against the JsonPlatform benchmark and it gives something around +3-5k RPS improvement for JSON. @benaadams thank you!