@davidfowl 2020-04-08
Safe and faster